+++ CHANGELOG.md
@@ -16,7 +16,7 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 * Defaults: defaults-actstracking, C++20 build environment to enable ACTS and its dependencies.
   Additional overrides setup to enable C++20 build compatibility, Fairship: added acts and hepmc3 as requirements,
   GEANT4: v10.7.3 -> v11.1.3, GEANT4VMC: v5-4 -> v6-1-p1, VMC: v1-1-p1 -> v2-0, pythia: v8230-ship -> v8311,
-  vgm: v5-2, evtGen: R01-06-00-ship -> R02-02-00-alice2, boost: 1.75.0 -> 1.83.0
+  vgm: v5-2, evtGen: R01-06-00-ship -> R02-02-00-alice2, boost: 1.75.0 -> 1.83.0, gcc: 11.5.0 -> 13.2.0
 
 ### Fixed
 
@@ -29,6 +29,8 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 * EvtGen: Added cmake build instructions required for R02-02-00
 * Fairship: Added build option to include acts (if found)
 * Fairship: Added acts and hepmc3 as dependency modules 
+* gcc-toolchain: Disabled gprofng in binutil config
+* gcc-toolchain: Rebuild mpfr, gmp to be used with gdb
 
 ## [25.08](https://github.com/ShipSoft/shipdist/tree/25.08)